Modify second method code by deleting extra class
